add metadata ingestion to unzip
